# Rework of all the compute_density methods