--- a/src/GeoParams.jl
+++ b/src/GeoParams.jl
@@ -93,7 +93,7 @@ export compute_conductivity,
         Set_TP_Conductivity
 
 using .MaterialParameters.LatentHeat
-export compute_latent_heat,
+export compute_latent_heat,compute_latent_heat!,
         ConstantLatentHeat
 
 using .MaterialParameters.RadioactiveHeat