**WARNING**: this PR contains code-breaking changes
`compute_density()` and `compute_density!()` have been reworked to work with an arbitrary number of input parameters. The argument corresponding to the density equation parameters is now a `NamedTuple` (i.e. `args=(;P=rand(20, 20)` or `args=(P=rand(20, 20), T = rand(20, 20)`; see `test/tes_Density.jl` for syntax examples).

Example of new usage to compute density with/without `ParallelStencil.jl`
```Julia
using GeoParams
using ParallelStencil
using ParallelStencil.FiniteDifferences2D
@init_parallel_stencil(Threads, Float64, 2)

# with ParallelStencil
@parallel_indices (i, j) function foo(rho, MatParam, Phases, args)
    rho[i,j] = compute_density(MatParam, Phases[i,j], ntuple_idx(args, i, j))
    return 
end

# standard Julia
function bar(rho, MatParam, Phases, args)
    for i in eachindex(rho)
        rho[i] = compute_density(MatParam, Phases[i], ntuple_idx(args, i))
    end
end

MatParam1 = Vector{AbstractMaterialParamsStruct}(undef, 4)
MatParam1[1] = SetMaterialParams(Name="Crust", Phase=0,
                            CreepLaws= (PowerlawViscous(), LinearViscous(η=1e23Pas)),
                            Density   = ConstantDensity(ρ=2900kg/m^3))
MatParam1[2] = SetMaterialParams(Name="Lower Crust", Phase=1,
                            CreepLaws= (PowerlawViscous(n=5.), LinearViscous(η=1e21Pas)),
                            Density  = Compressible_Density(ρ0=3000kg/m^3))
MatParam1[3] = SetMaterialParams(Name="Lower Crust", Phase=2,
                            CreepLaws= LinearViscous(η=1e21Pas),
                            Density  = ConstantDensity())
MatParam1[4] = SetMaterialParams(Name="Lower Crust", Phase=3,
                            CreepLaws= LinearViscous(η=1e21Pas),
                            Density  = ConstantDensity())
MatParam = Tuple(MatParam1)

Phases              = ones(Int64,400,400)*0;
Phases[:,20:end]   .= 1
Phases[:,200:end]  .= 2
Phases[:,300:end]  .= 3

rho     = zeros(size(Phases))
T       =  ones(size(Phases))
P       =  ones(size(Phases))*10
args = (P=P, T=T)

@parallel (1:size(Phases, 1), 1:size(Phases, 2)) foo(rho, MatParam, Phases, args) # ParallelStencil
bar(rho, MatParam, Phases, args) # standard Julia
```
@boriskaus It would be great if before merging you could test it out with your codes to see if it works fine or if there are any performance issues  (I have observed something like +20/40% speed in most of my tests).