@albert-de-montserrat 2022-04-09
Will need some extra work, didnt realise this breaks all the other `compute_something()` functions