@boriskaus 2022-04-10
On my machine, the new Julia version is actually faster than the existing routines:

**Previous:**
400x400:
```julia
julia> @btime compute_density!($rho, $MatParam, $Phases, $P,$T);
  929.500 μs (0 allocations: 0 bytes)
```
4000x4000:
```julia
julia> @btime compute_density!(rho, MatParam, Phases, P,T);
  104.833 ms (0 allocations: 0 bytes)
```

**This PR:**
400x400:
```julia
julia> @btime bar($rho, $MatParam, $Phases, $args) # standard Julia
  666.567 μs (0 allocations: 0 bytes)
```
4000x4000:
```julia
julia> @btime bar($rho, $MatParam, $Phases, $args) # standard Julia
  45.283 ms (0 allocations: 0 bytes)
```

One difference, though, is that `compute_density!` previously directly worked on arrays, whereas you here have to write your own loops over the arrays. From a user point of view I think we should provide such routines (perhaps by adding a multiple dispatch in case Phases is an array?). Similarly in case we use `ParallelStencil` (which was not yet implemented in the previous version, perhaps to be done with using [Requires.jl](https://github.com/JuliaPackaging/Requires.jl)?).