@albert-de-montserrat 2022-04-10
Cool! That's a much better performance boost than I expected :)  That `bar()` function is just a toy example for a one-to-one comparison with the syntax of a `ParallelStencil` kernel. All the pre-existing density methods still exist using the new code under the hood, for example [here](https://github.com/albert-de-montserrat/GeoParams.jl/blob/2d221ff8a639d3e983574b41ce7137193e709662/test/test_Density.jl#L166) for the in-place function using an integer array for the phase, and [here](https://github.com/albert-de-montserrat/GeoParams.jl/blob/2d221ff8a639d3e983574b41ce7137193e709662/test/test_Density.jl#L193) using phase ratios. I covered all the methods in `test_Density.jl` but if there's any missing functionality let me know and I can add it.

As for providing `ParallelStencil` kernels, I'm not sure what is the best approach. The problem with the current status of PS is that before parsing custom kernel with `@parallel` or `@parallel_indices` one needs to run `@init_parallel_stencil` to set up the device and dimension. Meaning that `Requires` will load our kernels as soon as `PS` is loaded, and crash. What about building a `GeoParams.ParallelStencilKernels` submodule that the user can optionally load after setting up `PS` ?

Btw I realised that the densities of type `PhaseDiagram_LookupTable` work, but are not type-stable. I didn't look in detail at what is the problem, but we should fix this before running it on GPUs.