@boriskaus 2022-04-10
Two additional use case are the following:

**1. Loop**
Typically we would want to use this within a timestep or iteration loop in which `P`,`T` changes. 
If I define the following:
```julia
function loop(rho,P,T,MatParam,Phases)
    for iter=1:10
        P .= rand(size(P))
        T .= rand(size(T))
        args = (P=P, T=T);
        bar(rho, MatParam, Phases, args) 
    end
    nothing
end
```
it indeed does not allocate: 
```julia
julia> @btime loop($P,$T,$rho,$MatParam, $Phases);
  667.065 ms (0 allocations: 0 bytes)
```

**2. Additive functions**
This is not the case for density, but for creep laws we often have several functions in parallel (say a dislocation and a diffusion creep law) and make a sum of the contribution of each (if using a strain rate decomposition). 
It is important to test that this is still non-allocating. As density only allows one input parameter [here](https://github.com/JuliaGeodynamics/GeoParams.jl/blob/49ec5e4793a43805ad3fab2f6c2dd2e134a04886/src/MaterialParameters.jl#L196), we would have to temporarily deactivate this.