@boriskaus 2022-04-10
> Btw I realised that the densities of type `PhaseDiagram_LookupTable` work, but are not type-stable. The problem is all the `::Any` annotations [here](https://github.com/albert-de-montserrat/GeoParams.jl/blob/2d221ff8a639d3e983574b41ce7137193e709662/src/PhaseDiagrams/PhaseDiagrams.jl#L18). Will all these paremeters always have random and different types?

No, most of them are `Float64`, so the structure can be made type stable. 
Moving the whole struct to the GPU is a different problem as we have strings as well (which are not actually required for computations on the GPU, but are mostly for info about the phase diagram). Moreover, we use the [Interpolations.jl](https://github.com/JuliaMath/Interpolations.jl) package to do the linear interpolation within the phase diagram. This does currently not work on the GPU (but I guess we can write a custom interpolation routine if this becomes a bottleneck).