@boriskaus 2022-04-20
This is a great PR, which makes it easier to deal with computational methods that require an arbitrary number of input fields. 

For now it works for all thermal and density parameters, but not yet for creep laws. Yet, we are merging now, so it can already be used in a real code, and such that it is also clear how new methods will have to be added.